Refactor GraphQL tests to improve readability and maintainability
Add tests for CopilotKit GraphQL queries and mutations.

* **GraphQL Queries Tests:**
  - Add `backend/tests/infrastructure-layer/test_copilotkit_graphql_queries.py` to test GraphQL queries for conversation management and chat functionality.
  - Include tests for querying user conversations, conversation details, and error handling.
  - Use descriptive test names and follow the AAA pattern.

* **GraphQL Mutations Tests:**
  - Add `backend/tests/infrastructure-layer/test_copilotkit_graphql_mutations.py` to test GraphQL mutations for conversation management and chat functionality.
  - Include tests for creating conversations, adding messages, and submitting chat requests.
  - Use descriptive test names and follow the AAA pattern.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/SPRIME01/GPTWrapperKit/tree/feature/copilotkit-integration?shareId=XXXX-XXXX-XXXX-XXXX).
